// src/zustand/auth-store/types.ts
export enum Types {
  AUTH_CHANGE_FIELDS = "AUTH_CHANGE_FIELDS",
  AUTH_CHANGE_REGISTER_FIELDS = "AUTH_CHANGE_REGISTER_FIELDS",

  AUTH_LOGIN_REQUEST = 'AUTH_LOGIN_REQUEST',
  AUTH_LOGIN_FULFILLED = 'AUTH_LOGIN_FULFILLED',
  AUTH_LOGIN_REJECTED = 'AUTH_LOGIN_REJECTED',

  AUTH_LOGOUT = "AUTH_LOGOUT",


  AUTH_REGISTER_REQUEST = 'AUTH_REGISTER_REQUEST',
  AUTH_REGISTER_FULFILLED = 'AUTH_REGISTER_FULFILLED',
  AUTH_REGISTER_REJECTED = 'AUTH_REGISTER_REJECTED',

}